Add Hangman web interface
