+++ .editorconfig
@@ -21,3 +21,8 @@ insert_final_newline = true
 indent_style = tab
 indent_size = 4
 insert_final_newline = true
+
+[*.html]
+indent_style = tab
+indent_size = 4
+insert_final_newline = true
